Create guess_game.py
In this guessing game, the player has to guess a random number between 1 and 9. They have 3 chances to make the correct guess. The code will provide feedback on whether the guessed number is higher or lower than the secret number. If the player guesses the correct number within the chances provided, they win; otherwise, they lose. The code has been improved by fixing the "loose" typo to "lose" and adjusting the number of chances to 3 in the user messages.